# openapi.yaml
openapi: 3.0.3
info:
  title: DivvyDoo API
  description: API for expense splitting and group payment management
  version: 1.0.0
  contact:
    name: DivvyDoo Team
servers:
  - url: http://localhost:8080/v1
    description: Local development server
security:
  - BearerAuth: []
...
paths:
...
  /users/{id}/expenses:
    get:
      tags:
        - Expenses
      summary: Get user's expenses
      description: Get all expenses for a user. Users can only access their own expenses.
      operationId: getUserExpenses
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Number of items to return (default 20)
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          required: false
          description: Number of items to skip (default 0)
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Expenses retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - cannot access other user's expenses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}/balances:
    get:
      tags:
        - Balances
      summary: Get user's balance summary
      description: Get balance summary for a user including total balance, group balances, and peer balances. Users can only access their own balances.
      operationId: getUserBalances
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '200':
          description: Balance summary retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBalanceSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - cannot access other user's balances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /groups:
    post:
      tags:
        - Groups
      summary: Create a new group
      description: Create a new expense sharing group. The creator becomes the admin.
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /groups/{id}:
    get:
      tags:
        - Groups
      summary: Get group details
      description: Get group details by group ID. User must be a member of the group.
      operationId: getGroup
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
      responses:
        '200':
          description: Group retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not a member of the group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /groups/{id}/leave:
    post:
      tags:
        - Groups
      summary: Leave a group
      description: Leave a group. User must be a member of the group.
      operationId: leaveGroup
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
      responses:
        '200':
          description: Left group successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not a member of the group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /groups/{id}/expenses:
    get:
      tags:
        - Expenses
      summary: Get group expenses
      description: Get all expenses for a group. User must be a member of the group.
      operationId: getGroupExpenses
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Number of items to return (default 20)
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          required: false
          description: Number of items to skip (default 0)
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Expenses retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not a member of the group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /groups/{id}/balances:
    get:
      tags:
        - Balances
      summary: Get group balances
      description: Get balance summary for all members of a group. User must be a member of the group.
      operationId: getGroupBalances
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID
          schema:
            type: string
      responses:
        '200':
          description: Balances retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserBalanceSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not a member of the group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /expenses:
    post:
      tags:
        - Expenses
      summary: Create a new expense
      description: Create a new expense. If group_id is provided, user must be a member of the group.
      operationId: createExpense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: Expense created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not a member of the group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /expenses/{id}:
    get:
      tags:
        - Expenses
      summary: Get expense details
      description: Get expense details by expense ID. User must be part of the expense.
      operationId: getExpense
      parameters:
        - name: id
          in: path
          required: true
          description: Expense ID
          schema:
            type: string
      responses:
        '200':
          description: Expense retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not part of the expense
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the login endpoint
  schemas:
...
    CreateGroupRequest:
      type: object
      required:
        - name
        - currency
      properties:
        name:
          type: string
          description: Group name
          example: Roommates
        currency:
          type: string
          description: Default currency for the group
          example: USD
...
    CreateExpenseRequest:
      type: object
      required:
        - title
        - amount
        - currency
        - paid_by
        - split
      properties:
        group_id:
          type: string
          description: Group ID (optional for personal expenses)
          example: grp_abc123
        title:
          type: string
          description: Expense title/description
          example: Dinner at restaurant
        amount:
          type: number
          format: double
          description: Total expense amount
          example: 100.50
        currency:
          type: string
          description: Currency code
          example: USD
        paid_by:
          type: array
          description: Users who paid for the expense
          items:
            $ref: '#/components/schemas/PaidByItem'
        split:
          $ref: '#/components/schemas/ExpenseSplit'
...
    Group:
      type: object
      properties:
        id:
          type: string
          description: MongoDB ObjectID
          example: 507f1f77bcf86cd799439011
        group_id:
          type: string
          description: Group ID
          example: grp_abc123
        name:
          type: string
          description: Group name
          example: Roommates
        members:
          type: array
          items:
            $ref: '#/components/schemas/GroupMember'
        currency:
          type: string
          description: Default currency
          example: USD
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        is_active:
          type: boolean
          description: Whether the group is active
          example: true
    GroupMember:
      type: object
      properties:
        user_id:
          type: string
          description: User ID
          example: usr_abc123
        role:
          type: string
          enum:
            - admin
            - member
          description: Member's role in the group
          example: member
        joined_at:
          type: string
          format: date-time
          description: When the member joined
        is_active:
          type: boolean
          description: Whether the member is active
          example: true
...
    Expense:
      type: object
      properties:
        id:
          type: string
          description: MongoDB ObjectID
          example: 507f1f77bcf86cd799439011
        expense_id:
          type: string
          description: Expense ID
          example: exp_abc123
        group_id:
          type: string
          description: Group ID (null for personal expenses)
          example: grp_abc123
        creator_id:
          type: string
          description: User ID of the expense creator
          example: usr_abc123
        title:
          type: string
          description: Expense title
          example: Dinner at restaurant
        amount:
          type: number
          format: double
          description: Total expense amount
          example: 100.50
        currency:
          type: string
          description: Currency code
          example: USD
        paid_by:
          type: array
          items:
            $ref: '#/components/schemas/PaidByItem'
        split:
          $ref: '#/components/schemas/ExpenseSplit'
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        is_deleted:
          type: boolean
          description: Whether the expense is deleted
          example: false
    PaidByItem:
      type: object
      required:
        - user_id
        - amount
      properties:
        user_id:
          type: string
          description: User ID of the payer
          example: usr_abc123
        amount:
          type: number
          format: double
          description: Amount paid by this user
          example: 50.25
    ExpenseSplit:
      type: object
      required:
        - type
        - details
      properties:
        type:
          type: string
          enum:
            - equal
            - exact
            - percentage
            - shares
          description: Split type
          example: equal
        details:
          type: array
          items:
            $ref: '#/components/schemas/SplitDetail'
    SplitDetail:
      type: object
      required:
        - user_id
        - value
      properties:
        user_id:
          type: string
          description: User ID
          example: usr_abc123
        value:
          type: number
          format: double
          description: Amount, percentage, or number of shares depending on split type
          example: 25.5
...
    UserBalanceSummary:
      type: object
      properties:
        user_id:
          type: string
          description: User ID
          example: usr_abc123
        total_balance:
          type: number
          format: double
          description: Total balance across all groups (positive = owed to you)
          example: 150.50
        group_balances:
          type: array
          items:
            $ref: '#/components/schemas/GroupBalance'
        peer_balances:
          type: array
          items:
            $ref: '#/components/schemas/PeerBalance'
        currency:
          type: string
          description: Currency code
          example: USD
        last_updated:
          type: string
          format: date-time
          description: Last balance update timestamp
    GroupBalance:
      type: object
      properties:
        group_id:
          type: string
          description: Group ID
          example: grp_abc123
        group_name:
          type: string
          description: Group name
          example: Roommates
        balance:
          type: number
          format: double
          description: Balance in this group (positive = owed to you)
          example: 75.25
    PeerBalance:
      type: object
      properties:
        peer_id:
          type: string
          description: Other user's ID
          example: usr_def456
        peer_name:
          type: string
          description: Other user's name
          example: Jane Smith
        balance:
          type: number
          format: double
          description: Balance with this peer (positive = they owe you, negative = you owe them)
          example: -25.00
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: Success message
          example: Operation completed successfully
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Invalid request body